Initial commit
This commit includes a CasADi Python API example of a differential
robot navigating to a target pose in the absence of obstacles. The
example runs a simulation with a reference pose and plots the states
and control actuation
throughout the entirety of the simulation. Moreover, it generates a
simple animation of the robot performing the manoeuvre.

The commit also includes a README with instruction to run the codes in
the repository and a LICENSE file containing the repository license.