Add CMake build
This commit adds CMake support to run examples using the CasADi C++ API.
The commit includes a number of tests to check the installation of the
CasADi C++ API and IPOPT with and without the HSL linear solvers.
Instructions to build the CasADi C++ API and the HSL linear solvers from
sources are also included in the updated README file.