Add CasADi C++ API example
This commit adds a CasADi C++ API example of an Ackermann-steered robot
navigating to a target waypoint within the prediction horizon. The
example plots the planned trajectory and the control acutation.

The commit includes a generalised C++ interface for the MPC controller
together with an example executable and an updated CMake file to build
said example. The commit also includes instructions on building the
Matplot++ library, required for generating the trajectory and control
actuation plots.